django-strawberry/requirements.txt
django
djangorestframework
strawberry-graphql[debug-server]